requirements.txt
amqp==2.4.2
billiard==3.5.0.5
boto3==1.9.115
botocore==1.12.115
cachelock==0.0.3
celery==4.2.1
docutils==0.14
jmespath==0.9.4
kombu==4.4.0
pycurl==7.43.0.2
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
